requirements.txt
fastapi
uvicorn
websockets
pydantic
aioconsole
termcolor
requests
motor
python-dotenv
passlib[bcrypt]
bcrypt
